rename project name | rename packages | change repository realization? now use Hibernate session and transaction
