+++ pom.xml
@@ -8,10 +8,10 @@
         <version>3.1.1</version>
     </parent>
     <groupId>com.example</groupId>
-    <artifactId>mate</artifactId>
+    <artifactId>springproj</artifactId>
     <version>1.0-SNAPSHOT</version>
     <packaging>jar</packaging>
-    <name>mate</name>
+    <name>springproj</name>
     <repositories>
         <repository>
             <id>central</id>
