--- src/main/java/com/example/repository/BookRepositoryImpl.java
@@ -1,28 +1,0 @@
-package com.example.repository;
-
-import com.example.model.Book;
-import jakarta.persistence.EntityManager;
-import jakarta.persistence.PersistenceContext;
-import java.util.List;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Repository;
-import org.springframework.transaction.annotation.Transactional;
-
-@Repository
-@RequiredArgsConstructor
-public class BookRepositoryImpl implements BookRepository {
-    @PersistenceContext
-    private EntityManager entityManager;
-
-    @Transactional
-    @Override
-    public Book save(Book book) {
-        entityManager.persist(book);
-        return book;
-    }
-
-    @Override
-    public List<Book> findAll() {
-        return entityManager.createQuery("SELECT b FROM Book b", Book.class).getResultList();
-    }
-}
