+++ src/main/resources/application.properties
@@ -1,4 +1,4 @@
-spring.datasource.url=jdbc:mysql://localhost:3306/mate_db
+spring.datasource.url=jdbc:mysql://localhost:3306/springproj_db
 spring.datasource.username=root
 spring.datasource.password=dfcz80509230175
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
