+++ src/test/java/dev/vasyl/proj/SpringprojApplicationTests.java
@@ -1,10 +1,10 @@
-package com.example;
+package dev.vasyl.proj;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-public class MateApplicationTests {
+public class SpringprojApplicationTests {
 
     @Test
     void contextLoads() {
